Add functionality to, create todo section
create todo section now assigns correct todo to the list,
assigns the correct todo to the currently selected list.

Bug noticed where if u add a todo to a specific list and go back and try
to add a todo to the currently selected list again. It will default back
to all current todos